refactor: restoreLatestSimulatoin dans le results-store
